chore(cdp): move BatchExportContext type to shared types file
